requirements.txt
# Exoplanet Detection AI - Requirements (Python 3.12 Compatible)

# Core Dependencies
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Plotting/Data Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# Authentication and Security
bcrypt>=4.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0

# OAuth Integration
authlib>=1.2.0
requests-oauthlib>=1.3.0

# Additional Dependencies
requests>=2.31.0
httpx>=0.28.0
Pillow>=10.0.0
openpyxl>=3.1.0

# Explainability Libraries
shap>=0.44.0
lime>=0.2.0.1

# Utilities
tqdm>=4.66.0